Refactor Glob -> Regex conversion using hash
Instead of making every value a lambda, I've only set the necessary
ones to lambdas.  We're doing a type check every time we check the
conversion value now; an alternative would be to make them all lambdas
(using `-> * {}` to ignore the args), and then we'd be making lambda
calls every time even when unnecessary.  I thought this way made the
CONVERSION_TABLE constant easier to parse.